requirements.txt
Click==7.0
pre-commit>=1.14.3
PyExifTool==0.1.1
